模块化; 引入 TypeScript 编译; 部分长 mapping 定义化
