+++ src/definitions/AreaCode.ts
@@ -1,0 +1,192 @@
+export default {
+    0: "不明",
+    135: "宗谷地方北部",
+    136: "宗谷地方南部",
+    125: "上川地方北部",
+    126: "上川地方中部",
+    127: "上川地方南部",
+    130: "留萌地方中北部",
+    131: "留萌地方南部",
+    139: "北海道利尻礼文",
+    150: "日高地方西部",
+    151: "日高地方中部",
+    152: "日高地方東部",
+    145: "胆振地方西部",
+    146: "胆振地方中東部",
+    110: "檜山地方",
+    105: "渡島地方北部",
+    106: "渡島地方東部",
+    107: "渡島地方西部",
+    140: "網走地方",
+    141: "北見地方",
+    142: "紋別地方",
+    165: "根室地方北部",
+    166: "根室地方中部",
+    167: "根室地方南部",
+    160: "釧路地方北部",
+    161: "釧路地方中南部",
+    155: "十勝地方北部",
+    156: "十勝地方中部",
+    157: "十勝地方南部",
+    119: "北海道奥尻島",
+    120: "空知地方北部",
+    121: "空知地方中部",
+    122: "空知地方南部",
+    100: "石狩地方北部",
+    101: "石狩地方中部",
+    102: "石狩地方南部",
+    115: "後志地方北部",
+    116: "後志地方東部",
+    117: "後志地方西部",
+    200: "青森県津軽北部",
+    201: "青森県津軽南部",
+    202: "青森県三八上北",
+    203: "青森県下北",
+    230: "秋田県沿岸北部",
+    231: "秋田県沿岸南部",
+    232: "秋田県内陸北部",
+    233: "秋田県内陸南部",
+    210: "岩手県沿岸北部",
+    211: "岩手県沿岸南部",
+    212: "岩手県内陸北部",
+    213: "岩手県内陸南部",
+    220: "宮城県北部",
+    221: "宮城県南部",
+    222: "宮城県中部",
+    240: "山形県庄内",
+    241: "山形県最上",
+    242: "山形県村山",
+    243: "山形県置賜",
+    250: "福島県中通り",
+    251: "福島県浜通り",
+    252: "福島県会津",
+    300: "茨城県北部",
+    301: "茨城県南部",
+    310: "栃木県北部",
+    311: "栃木県南部",
+    320: "群馬県北部",
+    321: "群馬県南部",
+    330: "埼玉県北部",
+    331: "埼玉県南部",
+    332: "埼玉県秩父",
+    350: "東京都23区",
+    351: "東京都多摩東部",
+    352: "東京都多摩西部",
+    354: "神津島",
+    355: "伊豆大島",
+    356: "新島",
+    357: "三宅島",
+    358: "八丈島",
+    359: "小笠原",
+    340: "千葉県北東部",
+    341: "千葉県北西部",
+    342: "千葉県南部",
+    360: "神奈川県東部",
+    361: "神奈川県西部",
+    420: "長野県北部",
+    421: "長野県中部",
+    422: "長野県南部",
+    410: "山梨県東部",
+    411: "山梨県中・西部",
+    412: "山梨県東部・富士五湖",
+    440: "静岡県伊豆",
+    441: "静岡県東部",
+    442: "静岡県中部",
+    443: "静岡県西部",
+    450: "愛知県東部",
+    451: "愛知県西部",
+    430: "岐阜県飛騨",
+    431: "岐阜県美濃東部",
+    432: "岐阜県美濃中西部",
+    460: "三重県北部",
+    461: "三重県中部",
+    462: "三重県南部",
+    370: "新潟県上越",
+    371: "新潟県中越",
+    372: "新潟県下越",
+    375: "新潟県佐渡",
+    380: "富山県東部",
+    381: "富山県西部",
+    390: "石川県能登",
+    391: "石川県加賀",
+    400: "福井県嶺北",
+    401: "福井県嶺南",
+    500: "滋賀県北部",
+    501: "滋賀県南部",
+    510: "京都府北部",
+    511: "京都府南部",
+    520: "大阪府北部",
+    521: "大阪府南部",
+    530: "兵庫県北部",
+    531: "兵庫県南東部",
+    532: "兵庫県南西部",
+    535: "兵庫県淡路島",
+    540: "奈良県",
+    550: "和歌山県北部",
+    551: "和歌山県南部",
+    580: "岡山県北部",
+    581: "岡山県南部",
+    590: "広島県北部",
+    591: "広島県南東部",
+    592: "広島県南西部",
+    570: "島根県東部",
+    571: "島根県西部",
+    575: "島根県隠岐",
+    560: "鳥取県東部",
+    562: "鳥取県中部",
+    563: "鳥取県西部",
+    600: "徳島県北部",
+    601: "徳島県南部",
+    610: "香川県東部",
+    611: "香川県西部",
+    620: "愛媛県東予",
+    621: "愛媛県中予",
+    622: "愛媛県南予",
+    630: "高知県東部",
+    631: "高知県中部",
+    632: "高知県西部",
+    700: "山口県北部",
+    702: "山口県西部",
+    703: "山口県東部",
+    704: "山口県中部",
+    710: "福岡県福岡",
+    711: "福岡県北九州",
+    712: "福岡県筑豊",
+    713: "福岡県筑後",
+    750: "大分県北部",
+    751: "大分県中部",
+    752: "大分県南部",
+    753: "大分県西部",
+    730: "長崎県北部",
+    731: "長崎県南西部",
+    732: "長崎県島原半島",
+    735: "長崎県対馬",
+    736: "長崎県壱岐",
+    737: "長崎県五島",
+    720: "佐賀県北部",
+    721: "佐賀県南部",
+    740: "熊本県阿蘇",
+    741: "熊本県熊本",
+    742: "熊本県球磨",
+    743: "熊本県天草・芦北",
+    760: "宮崎県北部平野部",
+    761: "宮崎県北部山沿い",
+    762: "宮崎県南部平野部",
+    763: "宮崎県南部山沿い",
+    770: "鹿児島県薩摩",
+    771: "鹿児島県大隅",
+    774: "鹿児島県十島村",
+    775: "鹿児島県甑島",
+    776: "鹿児島県種子島",
+    777: "鹿児島県屋久島",
+    778: "鹿児島県奄美北部",
+    779: "鹿児島県奄美南部",
+    800: "沖縄県本島北部",
+    801: "沖縄県本島中南部",
+    802: "沖縄県久米島",
+    803: "沖縄県大東島",
+    804: "沖縄県宮古島",
+    805: "沖縄県石垣島",
+    806: "沖縄県与那国島",
+    807: "沖縄県西表島"
+}
