+++ src/definitions/DrillType.ts
@@ -1,0 +1,8 @@
+export default {
+    "00": "通常",
+    "01": "訓練",
+    "10": "取り消し",
+    "11": "訓練取り消し",
+    "20": "参考情報またはテスト",
+    "30": "コード部のみの配信試験"
+}
