+++ src/definitions/TelegramType.ts
@@ -1,0 +1,6 @@
+export default {
+    "35": "最大予測震度のみの高度利用者向け緊急地震速報",
+    "36": "マグニチュード、最大予測震度及び主要動到達予測時刻の高度利用者向け緊急地震速報(B-Δ法、テリトリ法)",
+    "37": "マグニチュード、最大予測震度及び主要動到達予測時刻の高度利用者向け緊急地震速報(グリッドサーチ法、EPOS自動処理手法)",
+    "39": "キャンセル報"
+}
